+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change history for ui-quick-marc
 
-## [6.1.0](IN PROGRESS)
+## [7.0.0](IN PROGRESS)
 
 * [UIQM-250](https://issues.folio.org/browse/UIQM-250) Updated invalid Leader position values error messages to be more informative.
 * [UIQM-410](https://issues.folio.org/browse/UIQM-410) Add support for values 'b' and 'c' in Bib Leader/19
@@ -12,6 +12,7 @@
 * [UIQM-432](https://issues.folio.org/browse/UIQM-432) Avoid private paths in stripes-core imports.
 * [UIQM-421](https://issues.folio.org/browse/UIQM-421) Add correct validation to save derived records. 
 * [UIQM-445](https://issues.folio.org/browse/UIQM-445) Fix missing permission to view MARC Authority icon in MARC bib source view
+* [UIQM-447](https://issues.folio.org/browse/UIQM-447) Upgrade instance-authority-links interface to 2.0
 
 ## [6.0.2](https://github.com/folio-org/ui-quick-marc/tree/v6.0.2) (2023-03-30)